Use double for ElectrodeDataFrame::data (matches spec float64)
The SONATA spec defines scaling_factors as float64. Storing as double
avoids precision loss and removes the static_cast.

+++ include/bbp/sonata/electrode_reader.h
@@ -31,7 +31,7 @@ struct SONATA_API ElectrodeDataFrame {
     std::vector<uint64_t> electrodes;
 
     /// Flattened row-major data. data[row * n_cols + col] where n_cols = electrodes.size()
-    std::vector<float> data;
+    std::vector<double> data;
 };
 
 /**
